oyapassam

+++ 0x09-static_libraries/create_static_lib.sh
@@ -1,13 +1,20 @@
 #!/bin/bash
 
-/* Compile each .c file into a .o file */
-gcc -c -Wall -Werror -Wextra -pedantic -std=gnu89 *.c
+/* List of excluded files (e.g., main.c) */
+EXCLUDED_FILES="main.c"
 
-/* Create the static library liball.a from all .o files */
-ar -rc liball.a *.o
+/* Compile each .c file (excluding excluded files) into a .o file */
+for file in *.c; do
+	    if [[ ! " $EXCLUDED_FILES " =~ " $file " ]]; then
+		    		gcc -c -Wall -Werror -Wextra -pedantic -std=gnu89 "$file"
+					fi
+				done
 
-/* Index the library for faster linking */
-ranlib liball.a
+				/* Create the static library liball.a from all .o files (excluding excluded files) */
+				ar -rc liball.a *.o
 
-/* Clean up the .o files */
-rm *.o
+				/* Index the library for faster linking */
+				ranlib liball.a
+
+				/* Clean up the .o files */
+				rm *.o
